connect: carve accounts.rs (pure move)
Accounts and request authentication leave main.rs: session cookies, the
passkey register/login ceremonies, invites and the admin surface, the
per-request guards (bearer/daemon/rate-limit/origin/CSRF), handle
attestations, and the directory lookup. No unit tests reference these
items, so the tests module is untouched.
